requirements.txt
langchain = "^0.2.0"
langchain-openai = "^0.1.7"
svglib = "^1.5.1"
langchain-experimental = "^0.0.59"
cadquery = "^2.4.0"
rlpycairo = "^0.3.0"
loguru = "^0.7.2"
langchainhub = "^0.1.15"
python-dotenv = "^1.0.1"